+++ test/recording_studio_pages_test.rb
@@ -5,7 +5,7 @@ require "json"
 
 class RecordingStudioPagesTest < Minitest::Test
   def test_version_matches_release
-    assert_equal "0.3.5", ::RecordingStudioPages::VERSION
+    assert_equal "0.3.6", ::RecordingStudioPages::VERSION
   end
 
   def test_engine_exists